+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Features
+
+* **config:** add `auto_select_architect` option to automatically select swarm architect and disable competing built-in agents ([#887](https://github.com/zaxbysauce/opencode-swarm/issues/887))
+
 ## [7.21.5](https://github.com/zaxbysauce/opencode-swarm/compare/v7.21.4...v7.21.5) (2026-05-17)